FIX: fix dto data type

+++ app/src/main/java/com/example/rocketapp/model/rocket/info/RocketDiameter.kt
@@ -1,6 +1,6 @@
 package com.example.rocketapp.model.rocket.info
 
 data class RocketDiameter(
-    val meters: Int,
-    val feet: Int
+    val meters: Double,
+    val feet: Double
 )
