+++ app/src/main/java/com/example/rocketapp/model/rocket/stages/RocketSecondStage.kt
@@ -3,6 +3,6 @@ package com.example.rocketapp.model.rocket.stages
 data class RocketSecondStage(
     val reusable: Boolean,
     val engines: Int,
-    val fuelAmountTons: Int,
+    val fuelAmountTons: Double,
     val burnTimeSec: Int?
 )
